backend/requirements.txt
flask 
flask_cors 
boto3
numpy
scikit-learn
pillow
torch
torchvision
torchaudio
zappa